CopyPass improvements (mainly for multi-device usage)
Major changes:

- CopyPass: add a MultiDeviceCopyPass with just one InputOutput slot to copy within a multi-device buffer/image
- CopyPass: add support to copy depth/stencil images
- Add m_sourceFormat to CopyBufferToImageDescriptor to give more control and stay consistent with CopyImageToBufferDescriptor which has m_destinationFormat

Minor changes:

- Fix for CopyPass to create a transient attachment when cloneInput is enabled and the input is not transient.
- Add copy flags to be able to copy some buffers between devices (SkinnedMeshOutputStreamManager.cpp and SkinnedMeshVertexStreamProperties.cpp)
- Allow RW access when the scope attachment usage is copy as this happens when copying multi-device resources between devices (AttachmentEnums.cpp)
- Fix for BufferScopeAttachment during clear action error to actually check the write flag rather than exactly read write in line with the error message (BufferScopeAttachment.cpp)
- Add support for just depth OR stencil read only image layouts for subpass inputs (Conversion.cpp)

The change in Scope.cpp is a workaround for a specific issue we encountered where for a multi-aspect (depth and stencil) image with differing layouts for the aspects the layout transition is not done correctly. The situation is as follows:

1. We copy a depth stencil image (final layout: transfer dest)
2. We run a compute pass that reads only the depth image (final layout: depth read only optimal for depth, while stencil is still in transfer dest)
3. We run a render pass where we use it as actual depth stencil attachment

The Vulkan render pass needs a working initial layout and that is determined by the global barriers which are not actually submitted, since the render pass does the necessary layout transition. It can only do so if the layout of the aspect masks is the same or at least can be combined to a valid one (eg. depth read only and stencil read only to depth stencil read only). However with the two layouts the image has at the time it wants to combine them and the combination doesn't work because the stencil aspect is still in transfer dest and not in a attachment optimal or read only optimal layout. Therefore it is necessary to insert a barrier that transitions the stencil aspect to one of these which in our workaround is done by directly going to the target layout that the render pass would transition to. The combination then works because the (recursive) OptimizeBarrier call in the if inserts the necessary global barrier that the render pass uses to determine the initial layout combination. We then set `canOptimize` to false to actually insert a barrier in the scope barriers that gets executed for the actual transition from transfer dest to the target layout. Of course this is a very specific use case and there are likely other instances where something like this can happen for depth/stencil images (or other currently unsupported multi-aspect images), e.g., the other way around where depth is still in a transfer state and stencil is correct. The way barriers and layout transitions are currently handled doesn't allow for a nice solution here really unfortunately.

Co-authored-by: Joerg H. Mueller <joerg.mueller@huawei.com>
Co-authored-by: Martin Winter <martin.winter@huawei.com>
Signed-off-by: Joerg H. Mueller <joerg.mueller@huawei.com>